[#407] refactor: infoFix 디렉토리 이동
