+++ src/main/java/com/jeju/nanaland/domain/report/repository/InfoFixReportImageFileRepository.java
@@ -1,6 +1,6 @@
 package com.jeju.nanaland.domain.report.repository;
 
-import com.jeju.nanaland.domain.report.entity.InfoFixReportImageFile;
+import com.jeju.nanaland.domain.report.entity.infoFix.InfoFixReportImageFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InfoFixReportImageFileRepository extends
