+++ src/main/java/com/jeju/nanaland/domain/report/repository/InfoFixReportRepository.java
@@ -1,6 +1,6 @@
 package com.jeju.nanaland.domain.report.repository;
 
-import com.jeju.nanaland.domain.report.entity.InfoFixReport;
+import com.jeju.nanaland.domain.report.entity.infoFix.InfoFixReport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InfoFixReportRepository extends JpaRepository<InfoFixReport, Long> {
