+++ src/main/java/com/jeju/nanaland/domain/report/service/ReportService.java
@@ -25,8 +25,8 @@ import com.jeju.nanaland.domain.nature.repository.NatureRepository;
 import com.jeju.nanaland.domain.report.dto.ReportRequest;
 import com.jeju.nanaland.domain.report.dto.ReportRequest.ClaimReportDto;
 import com.jeju.nanaland.domain.report.dto.ReportRequest.InfoFixDto;
-import com.jeju.nanaland.domain.report.entity.FixType;
-import com.jeju.nanaland.domain.report.entity.InfoFixReport;
+import com.jeju.nanaland.domain.report.entity.infoFix.FixType;
+import com.jeju.nanaland.domain.report.entity.infoFix.InfoFixReport;
 import com.jeju.nanaland.domain.report.entity.Report;
 import com.jeju.nanaland.domain.report.entity.ReportStrategyFactory;
 import com.jeju.nanaland.domain.report.entity.ReportStrategy;
