+++ src/test/java/com/jeju/nanaland/domain/report/service/ReportServiceTest.java
@@ -32,10 +32,10 @@ import com.jeju.nanaland.domain.member.entity.Member;
 import com.jeju.nanaland.domain.member.entity.enums.TravelType;
 import com.jeju.nanaland.domain.member.repository.MemberRepository;
 import com.jeju.nanaland.domain.report.dto.ReportRequest;
-import com.jeju.nanaland.domain.report.entity.ClaimReportStrategy;
-import com.jeju.nanaland.domain.report.entity.FixType;
-import com.jeju.nanaland.domain.report.entity.InfoFixReport;
-import com.jeju.nanaland.domain.report.entity.InfoFixReportStrategy;
+import com.jeju.nanaland.domain.report.entity.claim.ClaimReportStrategy;
+import com.jeju.nanaland.domain.report.entity.infoFix.FixType;
+import com.jeju.nanaland.domain.report.entity.infoFix.InfoFixReport;
+import com.jeju.nanaland.domain.report.entity.infoFix.InfoFixReportStrategy;
 import com.jeju.nanaland.domain.report.entity.Report;
 import com.jeju.nanaland.domain.report.entity.ReportStrategyFactory;
 import com.jeju.nanaland.domain.report.entity.ReportType;
